debug

+++ Makefile
@@ -81,7 +81,15 @@ ifeq ($(OS),Windows_NT)
 $(CGO_INCLUDE_DIR)/$(CGO_OUTPUT_HEADER):
 	ls $(CGO_BUILD_DIR)
 	ls $(CGO_INCLUDE_DIR)
+	if [ ! -f $(CGO_BUILD_DIR)/$(CGO_OUTPUT_HEADER) ]; then
+		echo "Error: libwebrtc.h was not generated. Check the build configuration."
+		exit 1
+	fi	
 	mv $(CGO_BUILD_DIR)/$(CGO_OUTPUT_HEADER) $(CGO_INCLUDE_DIR)/$(CGO_OUTPUT_HEADER)
+	if [ ! -f $(CGO_INCLUDE_DIR)/$(CGO_OUTPUT_HEADER) ]; then
+		echo "Error: libwebrtc.h was not generated. Check the build configuration."
+		exit 1
+	fi		
 	powershell -ExecutionPolicy Bypass -File scripts\refine_header.ps1 -InputFile $(CGO_INCLUDE_DIR)/$(CGO_OUTPUT_HEADER) -OutputFile $(CGO_INCLUDE_DIR)/temp_header.h
 	mv /y $(CGO_INCLUDE_DIR)/temp_header.h $(CGO_INCLUDE_DIR)/$(CGO_OUTPUT_HEADER)
 	@echo Processed $(CGO_OUTPUT_HEADER): Removed problematic block.
